Added Select-AzureRmSubscription to automationAccountsScriptBlock
With multiple subscriptions I found that the Automation Accounts were being created in the tier 2 subscription. This was because when executing Add-AzureRmccount using my account it defaulted to the Tier2 Subscription. Since there is no long a way to set a default subscription (that I know of) it is pretty much mandatory to run Seletct-AzureRmSubscription after Add-AzureRmAccount.